admin-panel/requirements.txt
flask>=3.0.0
pyyaml>=6.0
feedparser>=6.0.0
